Move the palette's search suggestions out of command.store
Roughly 320 of the store's 746 lines were one command argument's getItems:
recents hydration and dedupe, frequents, curated categories, brand encoding,
GTFS stop resolution and the server autocomplete. That is a search concern, not
a command registry, and it kept the registry from being readable.

It becomes services/search-suggestions.service.ts. command.store.ts: 746 -> 436.
